+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 
 # Changelog
 
+## New Release
+
+### New features
+
+- Added `\Wizaplace\SDK\Company\AbstractCompanyRegistration::getInvoicingDisabled`
+- Added `\Wizaplace\SDK\Company\AbstractCompanyRegistration::setInvoicingDisabled`
+- Added `\Wizaplace\SDK\Company\Company::getInvoicingDisabled`
+- Added `\Wizaplace\SDK\Company\Company::setInvoicingDisabled`
+- Added `\Wizaplace\SDK\Company\CompanyUpdateCommand::getInvoicingDisabled`
+- Added `\Wizaplace\SDK\Company\CompanyUpdateCommand::setInvoicingDisabled`
+
 ## 1.128.0
 
 ### New features